+++ www/local-notification.js
@@ -29,6 +29,7 @@ exports._defaults = {
     autoClear     : true,
     badge         : null,
     channel       : null,
+    channelDescription: null,
     clock         : true,
     color         : null,
     data          : null,
@@ -56,7 +57,8 @@ exports._defaults = {
     title         : '',
     trigger       : { type : 'calendar' },
     vibrate       : false,
-    wakeup        : true
+    wakeup        : true,
+	when		  : 0
 };
 
 // Event listener